+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Change Log
 
+### Unreleased
+
 ### [v0.29](https://github.com/realityforge/proton/tree/v0.29) (2020-01-29) · [Full Changelog](https://github.com/realityforge/proton/compare/v0.28...v0.29)
 
 Changes in this release: