+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Change Log
 
-### Unreleased
+### [v0.30](https://github.com/realityforge/proton/tree/v0.30) (2020-01-29) · [Full Changelog](https://github.com/realityforge/proton/compare/v0.29...v0.30)
+
+Changes in this release:
 
 * Add `AbstractProcessorTest.input(...)` helper method.
 * Add `AbstractProcessorTest.assertCompilationUnsuccessful(...)` helper method.